[draft] move overlay to partial file
